Class "annoLinkerResult"

An object of class "annoLinkerResult" represent the annotated peaks, which is a GRanges object with peaks annotated by gene clusters, and interaction graph, which is an igraph graph.

# S4 method for class 'annoLinkerResult'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

anno_peaks(x)

# S4 method for class 'annoLinkerResult'
anno_peaks(x)

anno_graph(x)

# S4 method for class 'annoLinkerResult'
anno_graph(x)

anno_clusters(x)

# S4 method for class 'annoLinkerResult'
anno_clusters(x)

anno_evidence(x, i)

# S4 method for class 'annoLinkerResult'
anno_evidence(x, i)

anno_event(x, i)

# S4 method for class 'annoLinkerResult'
anno_event(x, i)

anno_feature(x, i)

# S4 method for class 'annoLinkerResult'
anno_feature(x, i)

anno_peakbin(x, i)

# S4 method for class 'annoLinkerResult'
anno_peakbin(x, i)

anno_featurebin(x, i)

# S4 method for class 'annoLinkerResult'
anno_featurebin(x, i)

# S4 method for class 'annoLinkerResult'
length(x)

# S4 method for class 'annoLinkerResult'
show(object)

# S4 method for class 'annoLinkerResult'
head(x, ...)

Arguments

x, object

An annoLinkerResult object.

row.names, optional, ...

parameters used by as.data.frame

i

Numeric, index value.

Objects from the Class

Objects can be created by calls of the form new("annoLinkerResult", annotated_peaks, graph, clusters).

Examples

library(igraph)
#> 
#> Attaching package: 'igraph'
#> The following objects are masked from 'package:rtracklayer':
#> 
#>     blocks, path
#> The following object is masked from 'package:GenomicRanges':
#> 
#>     union
#> The following object is masked from 'package:IRanges':
#> 
#>     union
#> The following object is masked from 'package:S4Vectors':
#> 
#>     union
#> The following objects are masked from 'package:BiocGenerics':
#> 
#>     normalize, path, union
#> The following objects are masked from 'package:generics':
#> 
#>     components, union
#> The following objects are masked from 'package:stats':
#> 
#>     decompose, spectrum
#> The following object is masked from 'package:base':
#> 
#>     union
library(GenomicRanges)
new("annoLinkerResult",
 annotated_peaks=GRanges(),
 graph=make_empty_graph(),
 clusters=data.frame())
#> An object of annoLinkerResult with annotated_peaks and interaction graph.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
#> IGRAPH 1815841 D--- 0 0 -- 
#> + edges from 1815841: